# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Ludo Game API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: number;
  username: string;
  email: string;
  coins: number;
  level: number;
  inviteCode: string;
  isAdmin: boolean;
  createdAt: string;
}

export interface UserRegistration {
  /** @minLength 3 */
  username: string;
  email: string;
  /** @minLength 6 */
  password: string;
  /** @nullable */
  inviteCode?: string | null;
}

export interface UserLogin {
  email: string;
  password: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface CoinBundle {
  id: number;
  coinsAmount: number;
  priceBdt: number;
  /** @nullable */
  originalPriceBdt: number | null;
  /** @nullable */
  discountPercent: number | null;
  label: string;
  isActive: boolean;
  sortOrder: number;
  createdAt: string;
}

export interface CoinBundleInput {
  coinsAmount: number;
  priceBdt: number;
  /** @nullable */
  originalPriceBdt?: number | null;
  /** @nullable */
  discountPercent?: number | null;
  label: string;
  sortOrder: number;
  isActive?: boolean;
}

export interface CoinBundleUpdate {
  coinsAmount?: number;
  priceBdt?: number;
  /** @nullable */
  originalPriceBdt?: number | null;
  /** @nullable */
  discountPercent?: number | null;
  label?: string;
  sortOrder?: number;
  isActive?: boolean;
}

export interface PurchaseInput {
  bundleId: number;
}

export interface PurchaseResult {
  coinsAdded: number;
  newBalance: number;
  bundleLabel: string;
}

export interface Notification {
  id: number;
  userId: number;
  title: string;
  message: string;
  notifType: string;
  isRead: boolean;
  createdAt: string;
}

export interface UnreadCount {
  count: number;
}

export interface RankingEntry {
  rank: number;
  userId: number;
  username: string;
  coins: number;
  level: number;
}

export interface InviteInfo {
  inviteCode: string;
  totalInvited: number;
  coinsEarned: number;
}

export interface Friend {
  id: number;
  username: string;
  coins: number;
  level: number;
}

export interface FriendInput {
  username: string;
}