# lib/api-zod/src/generated/types/coinBundleInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Ludo Game API
 * OpenAPI spec version: 0.2.0
 */

export interface CoinBundleInput {
  coinsAmount: number;
  priceBdt: number;
  /** @nullable */
  originalPriceBdt?: number | null;
  /** @nullable */
  discountPercent?: number | null;
  label: string;
  sortOrder: number;
  isActive?: boolean;
}