# lib/api-zod/src/generated/types/purchaseResult.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Ludo Game API
 * OpenAPI spec version: 0.2.0
 */

export interface PurchaseResult {
  coinsAdded: number;
  newBalance: number;
  bundleLabel: string;
}