# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Ludo Game API
 * OpenAPI spec version: 0.2.0
 */

export interface User {
  id: number;
  username: string;
  email: string;
  coins: number;
  level: number;
  inviteCode: string;
  isAdmin: boolean;
  createdAt: Date;
}