# lib/api-zod/src/generated/types/userRegistration.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Ludo Game API
 * OpenAPI spec version: 0.2.0
 */

export interface UserRegistration {
  /** @minLength 3 */
  username: string;
  email: string;
  /** @minLength 6 */
  password: string;
  /** @nullable */
  inviteCode?: string | null;
}